+++ mycodo/inputs/atlas_pt1000.py
@@ -1,6 +1,8 @@
 # coding=utf-8
 import copy
 
+from flask_babel import lazy_gettext
+
 from mycodo.inputs.base_input import AbstractInput
 from mycodo.utils.atlas_calibration import setup_atlas_device
 from mycodo.utils.system_pi import str_is_float
@@ -43,7 +45,27 @@ INPUT_INFORMATION = {
     'i2c_address_editable': True,
     'uart_location': '/dev/ttyAMA0',
     'uart_baud_rate': 9600,
-    'ftdi_location': '/dev/ttyUSB0'
+    'ftdi_location': '/dev/ttyUSB0',
+
+    'custom_actions_message':
+        'The I2C address can be changed. Enter a new address in the 0xYY format '
+        '(e.g. 0x22, 0x50), then press Set I2C Address. Remember to deactivate and '
+        'change the I2C address option after setting the new address.',
+
+    'custom_actions': [
+        {
+            'id': 'new_i2c_address',
+            'type': 'text',
+            'default_value': '0x67',
+            'name': lazy_gettext('New I2C Address'),
+            'phrase': lazy_gettext('The new I2C to set the device to')
+        },
+        {
+            'id': 'set_i2c_address',
+            'type': 'button',
+            'name': lazy_gettext('Set I2C Address')
+        }
+    ]
 }
 
 
@@ -114,3 +136,15 @@ class InputModule(AbstractInput):
         self.value_set(0, temp)
 
         return self.return_dict
+
+    def set_i2c_address(self, args_dict):
+        if 'new_i2c_address' not in args_dict:
+            self.logger.error("Cannot set new I2C address without an I2C address")
+            return
+        try:
+            i2c_address = int(str(args_dict['new_i2c_address']), 16)
+            write_cmd = "I2C,{}".format(i2c_address)
+            self.logger.debug("I2C Change command: {}".format(write_cmd))
+            self.atlas_device.write(write_cmd)
+        except:
+            self.logger.exception("Exception changing I2C address")