+++ mycodo/outputs/pump_atlas_ezo_pmp.py
@@ -100,6 +100,26 @@ OUTPUT_INFORMATION = {
             'name': lazy_gettext('Current (Amps)'),
             'phrase': lazy_gettext('The current draw of the device being controlled')
         }
+    ],
+
+    'custom_actions_message':
+        'The I2C address can be changed. Enter a new address in the 0xYY format '
+        '(e.g. 0x22, 0x50), then press Set I2C Address. Remember to deactivate and '
+        'change the I2C address option after setting the new address.',
+
+    'custom_actions': [
+        {
+            'id': 'new_i2c_address',
+            'type': 'text',
+            'default_value': '0x67',
+            'name': lazy_gettext('New I2C Address'),
+            'phrase': lazy_gettext('The new I2C to set the device to')
+        },
+        {
+            'id': 'set_i2c_address',
+            'type': 'button',
+            'name': lazy_gettext('Set I2C Address')
+        }
     ]
 }
 
@@ -252,3 +272,17 @@ class OutputModule(AbstractOutput):
         if self.atlas_command:
             return True
         return False
+
+    def set_i2c_address(self, args_dict):
+        self.output_setup = False
+
+        if 'new_i2c_address' not in args_dict:
+            self.logger.error("Cannot set new I2C address without an I2C address")
+            return
+        try:
+            i2c_address = int(str(args_dict['new_i2c_address']), 16)
+            write_cmd = "I2C,{}".format(i2c_address)
+            self.logger.debug("I2C Change command: {}".format(write_cmd))
+            self.atlas_command.write(write_cmd)
+        except:
+            self.logger.exception("Exception changing I2C address")